[FEAT] 시간표 생성기 서버 통신 수정
