+++ app/src/main/java/com/example/allclear/timetable/maketimetable/SelectSemesterActivity.java
@@ -6,14 +6,25 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.CompoundButton;
+import android.widget.Toast;
 
+import com.example.allclear.R;
+import com.example.allclear.data.ServicePool;
+import com.example.allclear.data.request.TimeTableOneRequestDto;
+import com.example.allclear.data.response.TimeTableOneResponseDto;
 import com.example.allclear.databinding.ActivitySelectSemesterBinding;
 import com.example.allclear.timetable.TimeTableFragment;
+
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 
 public class SelectSemesterActivity extends AppCompatActivity {
 
     private ActivitySelectSemesterBinding binding;
 
+    long userId = 1;
+    TimeTableOneRequestDto timeTableOneRequestDto;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -40,14 +51,32 @@ public class SelectSemesterActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 if(binding.cbEmptyTimeTable.isChecked()){
+                    // 빈 시간표 만드는 로직
                     Intent intent = new Intent(SelectSemesterActivity.this, TimeTableFragment.class);
                     startActivity(intent);
                 }else{
+                    postStepOneToServer(userId, timeTableOneRequestDto);
                     Intent intent = new Intent(SelectSemesterActivity.this, SelfAddOneActivity.class);
                     startActivity(intent);
                 }
             }
         });
+    }
+
+    private void postStepOneToServer(long userId, TimeTableOneRequestDto timeTableOneRequestDto) {
+        ServicePool.timeTableService.postStepOne(userId, timeTableOneRequestDto)
+                .enqueue(new Callback<TimeTableOneResponseDto>() {
+                    @Override
+                    public void onResponse(Call<TimeTableOneResponseDto> call, Response<TimeTableOneResponseDto> response) {
+                        // 선택한 학년, 학기 서버로 보내기
+                        Toast.makeText(SelectSemesterActivity.this, R.string.server_success, Toast.LENGTH_SHORT).show();
+                    }
+
+                    @Override
+                    public void onFailure(Call<TimeTableOneResponseDto> call, Throwable t) {
+                        Toast.makeText(SelectSemesterActivity.this, R.string.server_error, Toast.LENGTH_SHORT).show();
+                    }
+                });
     }
 
 }
