+++ app/src/main/res/layout/activity_select_semester.xml
@@ -12,7 +12,7 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="35dp"
         android:gravity="center"
-        android:text="시간표 만들기"
+        android:text="@string/add_time_table_title"
         android:textSize="28sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -35,7 +35,7 @@
         android:layout_height="wrap_content"
         android:layout_marginBottom="10dp"
         android:buttonTint="@color/first_blue"
-        android:text="빈 시간표 만들기"
+        android:text="@string/add_blank_time_table"
         android:textSize="12sp"
         app:layout_constraintBottom_toTopOf="@+id/btn_next"
         app:layout_constraintEnd_toEndOf="parent"
@@ -50,7 +50,7 @@
         android:backgroundTint="@color/first_blue"
         android:gravity="center"
         android:scaleType="fitCenter"
-        android:text="다음"
+        android:text="@string/btn_next"
         android:textColor="@color/white"
         android:textSize="16sp"
         app:layout_constraintBottom_toBottomOf="parent"
