+++ app/src/main/res/values/strings.xml
@@ -3,7 +3,12 @@
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="logo">올클</string>
-    <string name="add_timetable_title">시간표 만들기</string>
+
+    <string name="add_time_table_title">시간표 만들기</string>
+    <string name="add_blank_time_table">빈 시간표 만들기</string>
+
+    <string name="btn_next">다음</string>
+
     <string name="server_success">서버 통신 성공</string>
     <string name="server_error">서버 통신 실패</string>
 </resources>
